dataset_utils/pyproject.toml
[tool.poetry]
name = "dataset-utils"
version = "0.1.0"
description = ""
authors = ["Khiem Le <theinnocentman1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.9.0.80"
pillow = "^10.3.0"
imagesize = "^1.4.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
